requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.1.8
crud==0.1
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.121.3
h11==0.16.0
httptools==0.7.1
httpx==0.27.0
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1